Move triggerUpdate into listen_update.ts
