+++ src/web/entry/common.ts
@@ -4,11 +4,8 @@ import { UpdateCache } from './UpdateCache';
 import { WorkspaceConfig } from './WorkspaceConfig';
 
 const needUpdateSet: Set<string> = new Set();
-
 const updateCache = new UpdateCache();
-
 const updateInfo = new UpdateInfo();
-
 const workspaceConfig: WorkspaceConfig = new WorkspaceConfig();
 
-export {needUpdateSet, updateCache, updateInfo, workspaceConfig};
+export { needUpdateSet, updateCache, updateInfo, workspaceConfig };
